+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Calendar: Take holidays into account when viewing availability in participants tab
 - Support: task list grid bigger limit
 - Comments: If no labels available, disable 'Add label' button. Update tooltip.
+- Core: label color matches parent
 
 12-05-2023: 6.7.28
 - Core: Ioncube problems on PHP 8.1